{"schema":"decodex/commit/1","summary":"Split status queued attention projection","authority":"XY-1115"}
